+++ packages/orq-rc/src/orq_ai_sdk/models/guardrailruleupdateop.py
@@ -2,8 +2,11 @@
 
 from __future__ import annotations
 from .expression import Expression, ExpressionTypedDict
-from .expressioninput import ExpressionInput, ExpressionInputTypedDict
 from .guardrailref import GuardrailRef, GuardrailRefTypedDict
+from .optionalexpressioninput import (
+    OptionalExpressionInput,
+    OptionalExpressionInputTypedDict,
+)
 from datetime import datetime
 from orq_ai_sdk.types import (
     BaseModel,
@@ -23,7 +26,7 @@ class GuardrailRuleUpdateRequestBodyTypedDict(TypedDict):
     description: NotRequired[str]
     display_name: NotRequired[str]
     enabled: NotRequired[bool]
-    expression: NotRequired[ExpressionInputTypedDict]
+    expression: NotRequired[OptionalExpressionInputTypedDict]
     guardrails: NotRequired[List[GuardrailRefTypedDict]]
     timeout: NotRequired[int]
 
@@ -35,7 +38,7 @@ class GuardrailRuleUpdateRequestBody(BaseModel):
 
     enabled: Optional[bool] = None
 
-    expression: Optional[ExpressionInput] = None
+    expression: Optional[OptionalExpressionInput] = None
 
     guardrails: Optional[List[GuardrailRef]] = None