+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPNJFE3SDT978KJ90EG6WJC7"
+        "01KPNT9JPRV3QEX92YEC57AYS3"
     )
     r"""The unique identifier of the data source"""