+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPNJFE3YGX920XR78ZWDHEJ5"
+        "01KPNT9JPX7MWBB9WS1GN4N6GD"
     )
     r"""The unique identifier of the data source"""